+++ app/DataTables/QuotesTrashDataTable.php
@@ -62,7 +62,7 @@ class QuotesTrashDataTable extends QuotesDataTable
      */
     public function query(Quote $model)
     {
-        return $model->has('client')->with('client')->where('invoice_id', 0)->onlyTrashed();
+        return $model->has('client')->with('client')->onlyTrashed();
     }
 
     /**