+++ app/Observers/InvoiceObserver.php
@@ -9,6 +9,7 @@ use BT\Modules\Groups\Models\Group;
 use BT\Modules\Invoices\Models\Invoice;
 use BT\Modules\Invoices\Support\InvoiceCalculate;
 use BT\Modules\Quotes\Models\Quote;
+use BT\Modules\Workorders\Models\Workorder;
 use BT\Support\DateFormatter;
 use BT\Support\Statuses\InvoiceStatuses;
 
@@ -20,10 +21,11 @@ class InvoiceObserver
     {
         $this->invoiceCalculate = $invoiceCalculate;
     }
+
     /**
      * Handle the invoice "created" event.
      *
-     * @param  \BT\Modules\Invoices\Models\Invoice  $invoice
+     * @param \BT\Modules\Invoices\Models\Invoice $invoice
      * @return void
      */
     public function created(Invoice $invoice): void
@@ -41,78 +43,63 @@ class InvoiceObserver
     /**
      * Handle the invoice "creating" event.
      *
-     * @param  \BT\Modules\Invoices\Models\Invoice  $invoice
+     * @param \BT\Modules\Invoices\Models\Invoice $invoice
      * @return void
      */
     public function creating(Invoice $invoice): void
     {
-        if (!$invoice->client_id)
-        {
+        if (!$invoice->client_id) {
             // This needs to throw an exception since this is required.
         }
 
-        if (!$invoice->user_id)
-        {
+        if (!$invoice->user_id) {
             $invoice->user_id = auth()->user()->id;
         }
 
-        if (!$invoice->invoice_date)
-        {
+        if (!$invoice->invoice_date) {
             $invoice->invoice_date = date('Y-m-d');
         }
 
-        if (!$invoice->due_at)
-        {
+        if (!$invoice->due_at) {
             $invoice->due_at = DateFormatter::incrementDateByDays($invoice->invoice_date->format('Y-m-d'), $invoice->client->client_terms);
         }
 
-        if (!$invoice->company_profile_id)
-        {
+        if (!$invoice->company_profile_id) {
             $invoice->company_profile_id = config('bt.defaultCompanyProfile');
         }
 
-        if (!$invoice->group_id)
-        {
+        if (!$invoice->group_id) {
             $invoice->group_id = config('bt.invoiceGroup');
         }
 
-        if (!$invoice->number)
-        {
+        if (!$invoice->number) {
             $invoice->number = Group::generateNumber($invoice->group_id);
         }
 
-        if (!isset($invoice->terms))
-        {
+        if (!isset($invoice->terms)) {
             $invoice->terms = config('bt.invoiceTerms');
         }
 
-        if (!isset($invoice->footer))
-        {
+        if (!isset($invoice->footer)) {
             $invoice->footer = config('bt.invoiceFooter');
         }
 
-        if (!$invoice->invoice_status_id)
-        {
+        if (!$invoice->invoice_status_id) {
             $invoice->invoice_status_id = InvoiceStatuses::getStatusId('draft');
         }
 
-        if (!$invoice->currency_code)
-        {
+        if (!$invoice->currency_code) {
             $invoice->currency_code = $invoice->client->currency_code;
         }
 
-        if (!$invoice->template)
-        {
+        if (!$invoice->template) {
             $invoice->template = $invoice->companyProfile->invoice_template;
         }
 
-        if ($invoice->currency_code == config('bt.baseCurrency'))
-        {
+        if ($invoice->currency_code == config('bt.baseCurrency')) {
             $invoice->exchange_rate = 1;
-        }
-        elseif (!$invoice->exchange_rate)
-        {
-            $currencyConverter      = CurrencyConverterFactory::create();
+        } elseif (!$invoice->exchange_rate) {
+            $currencyConverter = CurrencyConverterFactory::create();
             $invoice->exchange_rate = $currencyConverter->convert(config('bt.baseCurrency'), $invoice->currency_code);
         }
 
@@ -120,46 +107,82 @@ class InvoiceObserver
     }
 
     /**
-     * Handle the invoice "deleted" event.
+     * Handle the invoice "deleting" event.
      *
-     * @param  \BT\Modules\Invoices\Models\Invoice  $invoice
+     * @param \BT\Modules\Invoices\Models\Invoice $invoice
      * @return void
      */
-    public function deleteing(Invoice $invoice): void
+    public function deleting(Invoice $invoice): void
     {
-        foreach ($invoice->activities as $activity)
-        {
+        foreach ($invoice->activities as $activity) {
             ($invoice->isForceDeleting()) ? $activity->onlyTrashed()->forceDelete() : $activity->delete();
         }
 
-        foreach ($invoice->attachments as $attachment)
-        {
+        foreach ($invoice->attachments as $attachment) {
             ($invoice->isForceDeleting()) ? $attachment->onlyTrashed()->forceDelete() : $attachment->delete();
         }
 
-        foreach ($invoice->mailQueue as $mailQueue)
-        {
+        foreach ($invoice->mailQueue as $mailQueue) {
             ($invoice->isForceDeleting()) ? $mailQueue->onlyTrashed()->forceDelete() : $mailQueue->delete();
         }
 
-        foreach ($invoice->notes as $note)
-        {
+        foreach ($invoice->notes as $note) {
             ($invoice->isForceDeleting()) ? $note->onlyTrashed()->forceDelete() : $note->delete();
         }
 
-        Quote::where('invoice_id', $invoice->id)->update(['invoice_id' => 0]);
+        // set invoice_id ref in quote, workorder and expense to negative, denoting trashed
+        if ($invoice->quote() && !$invoice->isForceDeleting()) $invoice->quote()->update(['invoice_id' => -($invoice->id)]);
 
-        Expense::where('invoice_id', $invoice->id)->update(['invoice_id' => 0]);
+        if ($invoice->workorder() && !$invoice->isForceDeleting()) $invoice->workorder()->update(['invoice_id' => -($invoice->id)]);
 
-        $group = Group::where('id', $invoice->group_id)
-            ->where('last_number', $invoice->number)
-            ->first();
+        if ($invoice->expense() && !$invoice->isForceDeleting()) $invoice->expense()->update(['invoice_id' => -($invoice->id)]);
 
-        if ($group)
-        {
-            $group->next_id = $group->next_id - 1;
-            $group->save();
-        }
+        // todo this gets messy with soft deletes...
+//        $group = Group::where('id', $invoice->group_id)
+//            ->where('last_number', $invoice->number)
+//            ->first();
+//
+//        if ($group) {
+//            $group->next_id = $group->next_id - 1;
+//            $group->save();
+//        }
     }
+
+    /**
+     * Handle the invoice "restoring" event.
+     *
+     * @param \BT\Modules\Invoices\Models\Invoice $invoice
+     * @return void
+     */
+    public function restoring(Invoice $invoice): void
+    {
+        foreach ($invoice->activities as $activity) {
+            $activity->onlyTrashed()->restore();
+        }
+
+        foreach ($invoice->attachments as $attachment) {
+            $attachment->onlyTrashed()->restore();
+        }
+
+        foreach ($invoice->mailQueue as $mailQueue) {
+            $mailQueue->onlyTrashed()->restore();
+        }
+
+        foreach ($invoice->notes as $note) {
+            $note->onlyTrashed()->restore();
+        }
+
+        // if exists, remove negative to denote restored
+        $quote = Quote::where('invoice_id', -($invoice->id))->first();
+        if ($quote) $quote->update(['invoice_id' => $invoice->id]);
+
+        $workorder = Workorder::where('invoice_id', -($invoice->id))->first();
+        if ($workorder) $workorder->update(['invoice_id' => $invoice->id]);
+
+        $expense = Expense::where('invoice_id', -($invoice->id))->first();
+        if ($expense) $expense->update(['invoice_id' => $invoice->id]);
+
+    }
+
 
 }