+++ app/Observers/WorkorderObserver.php
@@ -6,6 +6,7 @@ namespace BT\Observers;
 use BT\Modules\Currencies\Support\CurrencyConverterFactory;
 use BT\Modules\CustomFields\Models\WorkorderCustom;
 use BT\Modules\Groups\Models\Group;
+use BT\Modules\Quotes\Models\Quote;
 use BT\Modules\Workorders\Models\Workorder;
 use BT\Modules\Workorders\Support\WorkorderCalculate;
 use BT\Support\DateFormatter;
@@ -21,7 +22,7 @@ class WorkorderObserver
     /**
      * Handle the workorder "created" event.
      *
-     * @param  \BT\Modules\Workorders\Models\Workorder  $workorder
+     * @param \BT\Modules\Workorders\Models\Workorder $workorder
      * @return void
      */
     public function created(Workorder $workorder): void
@@ -39,93 +40,75 @@ class WorkorderObserver
     /**
      * Handle the workorder "creating" event.
      *
-     * @param  \BT\Modules\Workorders\Models\Workorder  $workorder
+     * @param \BT\Modules\Workorders\Models\Workorder $workorder
      * @return void
      */
     public function creating(Workorder $workorder): void
     {
-        if (!$workorder->client_id)
-        {
+        if (!$workorder->client_id) {
             // This needs to throw an exception since this is required.
         }
 
-        if (!$workorder->user_id)
-        {
+        if (!$workorder->user_id) {
             $workorder->user_id = auth()->user()->id;
         }
 
-        if (!$workorder->workorder_date)
-        {
+        if (!$workorder->workorder_date) {
             $workorder->workorder_date = date('Y-m-d');
         }
 
-        if (!$workorder->job_date)
-        {
+        if (!$workorder->job_date) {
             $workorder->job_date = date('Y-m-d');
         }
 
-        if (!$workorder->start_time)
-        {
+        if (!$workorder->start_time) {
             $workorder->start_time = '08:00';
         }
 
-        if (!$workorder->end_time)
-        {
+        if (!$workorder->end_time) {
             $workorder->end_time = '09:00';
         }
 
-        if (!$workorder->expires_at)
-        {
+        if (!$workorder->expires_at) {
             $workorder->expires_at = DateFormatter::incrementDateByDays($workorder->workorder_date->format('Y-m-d'), config('bt.workordersExpireAfter'));
         }
 
-        if (!$workorder->company_profile_id)
-        {
+        if (!$workorder->company_profile_id) {
             $workorder->company_profile_id = config('bt.defaultCompanyProfile');
         }
 
-        if (!$workorder->group_id)
-        {
+        if (!$workorder->group_id) {
             $workorder->group_id = config('bt.workorderGroup');
         }
 
-        if (!$workorder->number)
-        {
+        if (!$workorder->number) {
             $workorder->number = Group::generateNumber($workorder->group_id);
         }
 
-        if (!isset($workorder->terms))
-        {
+        if (!isset($workorder->terms)) {
             $workorder->terms = config('bt.workorderTerms');
         }
 
-        if (!isset($workorder->footer))
-        {
+        if (!isset($workorder->footer)) {
             $workorder->footer = config('bt.workorderFooter');
         }
 
-        if (!$workorder->workorder_status_id)
-        {
+        if (!$workorder->workorder_status_id) {
             $workorder->workorder_status_id = WorkorderStatuses::getStatusId('draft');
         }
 
-        if (!$workorder->currency_code)
-        {
+        if (!$workorder->currency_code) {
             $workorder->currency_code = $workorder->client->currency_code;
         }
 
-        if (!$workorder->template)
-        {
+        if (!$workorder->template) {
             $workorder->template = config('bt.workorderTemplate');
         }
 
-        if ($workorder->currency_code == config('bt.baseCurrency'))
-        {
+        if ($workorder->currency_code == config('bt.baseCurrency')) {
             $workorder->exchange_rate = 1;
-        }
-        elseif (!$workorder->exchange_rate)
-        {
-            $currencyConverter    = CurrencyConverterFactory::create();
+        } elseif (!$workorder->exchange_rate) {
+            $currencyConverter = CurrencyConverterFactory::create();
             $workorder->exchange_rate = $currencyConverter->convert(config('bt.baseCurrency'), $workorder->currency_code);
         }
 
@@ -135,43 +118,70 @@ class WorkorderObserver
 
 
     /**
-     * Handle the workorder "deleted" event.
+     * Handle the workorder "deleting" event.
      *
-     * @param  \BT\Modules\Workorders\Models\Workorder  $workorder
+     * @param \BT\Modules\Workorders\Models\Workorder $workorder
      * @return void
      */
-    public function deleteing(Workorder $workorder): void
+    public function deleting(Workorder $workorder): void
     {
-        foreach ($workorder->activities as $activity)
-        {
+        foreach ($workorder->activities as $activity) {
             ($workorder->isForceDeleting()) ? $activity->onlyTrashed()->forceDelete() : $activity->delete();
         }
 
-        foreach ($workorder->attachments as $attachment)
-        {
+        foreach ($workorder->attachments as $attachment) {
             ($workorder->isForceDeleting()) ? $attachment->onlyTrashed()->forceDelete() : $attachment->delete();
         }
 
-        foreach ($workorder->mailQueue as $mailQueue)
-        {
+        foreach ($workorder->mailQueue as $mailQueue) {
             ($workorder->isForceDeleting()) ? $mailQueue->onlyTrashed()->forceDelete() : $mailQueue->delete();
         }
 
-        foreach ($workorder->notes as $note)
-        {
+        foreach ($workorder->notes as $note) {
             ($workorder->isForceDeleting()) ? $note->onlyTrashed()->forceDelete() : $note->delete();
         }
 
-        $group = Group::where('id', $workorder->group_id)
-            ->where('last_number', $workorder->number)
-            ->first();
+        //set workorder_id ref in quote to negative, denoting trashed
+        if ($workorder->quote() && !$workorder->isForceDeleting()) $workorder->quote()->update(['workorder_id' => -($workorder->id)]);
 
-        if ($group)
-        {
-            $group->next_id = $group->next_id - 1;
-            $group->save();
-        }
+        // todo this gets messy with soft deletes...
+//        $group = Group::where('id', $workorder->group_id)
+//            ->where('last_number', $workorder->number)
+//            ->first();
+//
+//        if ($group)
+//        {
+//            $group->next_id = $group->next_id - 1;
+//            $group->save();
+//        }
     }
 
+    /**
+     * Handle the workorder "restoring" event.
+     *
+     * @param \BT\Modules\Workorders\Models\Workorder $workorder
+     * @return void
+     */
+    public function restoring(Workorder $workorder): void
+    {
+        foreach ($workorder->activities as $activity) {
+            $activity->onlyTrashed()->restore();
+        }
 
+        foreach ($workorder->attachments as $attachment) {
+            $attachment->onlyTrashed()->restore();
+        }
+
+        foreach ($workorder->mailQueue as $mailQueue) {
+            $mailQueue->onlyTrashed()->restore();
+        }
+
+        foreach ($workorder->notes as $note) {
+            $note->onlyTrashed()->restore();
+        }
+
+        // if exists, remove negative to denote restored
+        $quote = Quote::where('workorder_id', -($workorder->id))->first();
+        if ($quote) $quote->update(['workorder_id' => $workorder->id]);
+    }
 }